// src/modules/sot/common/sot.enums.ts
export enum SotListFilter {
  NEAR_BY = 'nearby',
  SORTED = 'sorted',
}

export enum MapSotListFilter {
  ALL = 'all',
  ONGOING_SURVEY = 'ongoing',
  AVAILABLE_FOR_SURVEY = 'available',
}

export enum SotData {
  COUNTRY = 'Country',
  CITY = 'City',
  GRADE = 'Grade',
  UUID = 'UUID',
  LONGITUDE = 'Longitude',
  LATITUDE = 'Latitude',
}

export enum BlockInfo {
  HEIGHT = 'height',
}

export enum BlockQueue {
  BLOCK = 'block',
}

export enum JobAttempts {
  THREE = 3,
}

export enum JobEvents {
  Completed = 'completed',
}

export enum EarthInfo {
  RADIUS = 6371,
}

export enum DistanceKiloMeters {
  ONE = 1,
}