// src/modules/survey/common/survey.enums.ts
export enum SurveyType {
  SINGLE = 'single',
  MULTIPLE = 'multiple',
  SILVERBELL = 'silverbell',
}

export enum SurveyStatus {
  DISABLED = 'disabled',
  ENABLED = 'enabled',
  ONGOING = 'ongoing',
  ENDED = 'ended',
}

export enum SilverBellRequestStatus {
  PENDING = 'pending',
  APPROVED = 'approved',
  DECLINED = 'declined',
  IN_PROGRESS = 'in_progress',
}

export enum SurveyParticipationStatus {
  ONGOING = 'ongoing',
  ENDED = 'ended',
}

export enum Sort {
  ASCENDING = 'asc',
  DESCENDING = 'desc',
}

export enum CommentOrderBy {
  ASCENDING = 'asc',
  DESCENDING = 'desc',
  LIKES = 'likes',
}

export enum AgeGroupEnum {
  TEN = 10,
  TWENTY = 20,
  THIRTY = 30,
  FOURTY = 40,
  FIFTY = 50,
  SIXTY = 60,
}